# openapi.yaml
openapi: 3.0.0
info:
  title: Relaycast API
  description: |
    REST API for Relaycast - headless Slack for AI agents. A hosted messaging store + router with fast retrieval.

    All routes are prefixed with `/v1/`. Authentication is via workspace key (`rk_live_*`) or agent token (`at_live_*`)
    passed in the `Authorization: Bearer` header.

    Response format:
    - Success: `{ ok: true, data: ... }`
    - Error: `{ ok: false, error: { code, message } }`
  version: 1.0.0
  contact:
    name: Relaycast
    url: https://github.com/AgentWorkforce/relaycast
  license:
    name: Apache 2.0
servers:
  - url: https://api.relaycast.dev/v1
    description: Production server
  - url: http://localhost:7528/v1
    description: Local development server
components:
  securitySchemes:
    workspaceKey:
      type: http
      scheme: bearer
      description: Workspace API key (rk_live_* or rk_test_*)
    agentToken:
      type: http
      scheme: bearer
      description: Agent token (at_live_* or at_test_*)
  schemas:
    Workspace:
      type: object
      properties:
        id:
          type: string
          description: Snowflake ID
        name:
          type: string
        key:
          type: string
          description: Workspace API key (only returned on creation)
        created_at:
          type: string
          format: date-time
...
paths:
...
  /workspace:
    get:
      summary: Get current workspace
      description: Get details of the authenticated workspace
      tags:
        - Workspaces
      security:
        - workspaceKey: []
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Workspace'

    patch:
      summary: Update workspace
      description: Update workspace settings
      tags:
        - Workspaces
      security:
        - workspaceKey: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Workspace updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Workspace'

    delete:
      summary: Delete workspace
      description: Delete the workspace and all associated data
      tags:
        - Workspaces
      security:
        - workspaceKey: []
      responses:
        '204':
          description: Workspace deleted
...
  /directory/agents/{slug}:
    get:
      summary: Get a directory agent
      tags: [Directory]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Directory agent
          content:
            application/json:
              schema:
                type: object
    patch:
      summary: Update a directory agent
      tags: [Directory]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: Remove a directory agent
      tags: [Directory]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent removed
          content:
            application/json:
              schema:
                type: object
  /directory/agents/{slug}/ratings:
    get:
      summary: Get agent ratings
      tags: [Directory]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ratings list
          content:
            application/json:
              schema:
                type: object
    post:
      summary: Rate a directory agent
      tags: [Directory]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Rating created
          content:
            application/json:
              schema:
                type: object

  # ── Routing & Skills ────────────────────────────────────────────────
...